backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sympy==1.12
openai==1.3.5
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.0.3
